Solved rounding errors, lead to discover that equation was getting infinitely big and eventually, just returned 0 for super fun times. Changed how the crossover function works it now is limited to 2 times the max number of operators

+++ GeneticProgram/src/main/java/com/genetic/program/tree/BinaryMathTree.java
@@ -1,5 +1,6 @@
 package com.genetic.program.tree;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -20,7 +21,7 @@ public class BinaryMathTree {
 		this.rootNode = rootNode;
 	}
 
-	public double getValue(HashMap<String, Double> variables){
+	public BigDecimal getValue(HashMap<String, BigDecimal> variables){
 		return rootNode.getValue(variables);
 	}
